requirements.txt
django
djangorestframework
psycopg2
drf-spectacular
pillow